requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
Django==5.1.2
filelock==3.16.1
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.4
jiter==0.6.1
joblib==1.4.2
llama-cloud==0.1.4
llama-index==0.11.18
llama-index-agent-openai==0.3.4
llama-index-cli==0.3.1
llama-index-core==0.11.18
llama-index-embeddings-huggingface==0.3.1
llama-index-embeddings-openai==0.2.5
llama-index-indices-managed-llama-cloud==0.4.0
llama-index-legacy==0.9.48.post3
llama-index-llms-groq==0.2.0
llama-index-llms-openai==0.2.15
llama-index-llms-openai-like==0.2.0
llama-index-llms-replicate==0.2.1
llama-index-multi-modal-llms-openai==0.2.2
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.2
llama-index-readers-llama-parse==0.3.0
llama-parse==0.5.10
MarkupSafe==3.0.1
marshmallow==3.23.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.1
nltk==3.9.1
numpy==1.26.4
openai==1.52.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
propcache==0.2.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyOpenSSL==24.2.1
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
replicate==1.0.2
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.2.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.1
striprtf==0.0.26
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.1
torch==2.5.0
tqdm==4.66.5
transformers==4.45.2
Twisted==24.7.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wrapt==1.16.0
yarl==1.15.4
zope.interface==7.1.0
